+++ session.go
@@ -503,8 +503,10 @@ func (s *session) Context() context.Context {
 	return s.ctx
 }
 
-func (s *session) ConnectionState() tls.ConnectionState {
-	return s.cryptoStreamHandler.ConnectionState()
+func (s *session) ConnectionState() ConnectionState {
+	return ConnectionState{
+		TLS: s.cryptoStreamHandler.ConnectionState(),
+	}
 }
 
 func (s *session) maybeResetTimer() {